PrimeraPreentregaPF/sistema_ventas_comestibles/requirements.txt
mysql-connector-python==8.2.0
pytest==7.4.3
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3